Flesh out derive macro for Encode (#396)
This picks up where #324 left off, and is based on that branch. I've
opened a separate PR to aid review assuming a reader is already familiar
with that one.

At a high level, this makes the following changes.

0. Rename Loggable→Encode
1. Support for additional protobuf field types (ints, bytes, etc); see
protobuf.rs
2. Fix support for &str with lifetimes
3. Support generic structs
4. Do not panic if writing a large value (string, bytes) would overflow
the buffer

I also re-added documentation (and an example) for deriving JSON
schemas. I think this is still valuable for some use cases; for example,
the `mcap` CLI supports `--json` formatting on messages.

Not addressed here, but could consider in a future PR:

- Putting this behind a "derive" feature
- Some mechanism to override the protobuf package or schema name
(perhaps an attr macro?)

Note we will need to come up with a viable release plan prior to merging
this feature. I think the derive crate will need to be published first,
and awaited on, before we continue to publish the primary crate.