Add encoding_len to ProtobufField (#408)
This builds on #324, and is a follow-up to #396.

The primary change is to add a required `encoded_len` method to the
ProtobufField trait. This provides the length of encoded data to
`Encode::encode`, which will now return an error if the remaining buffer
is too short to hold the written data.

This also adds a compiler check for unit-only enums; any variant with
fields will now provide a nicer error message.

This also updates the documentation, and notes the protobuf limitation
around enum values.

This also fixes the encoding of tags, which were previously written as a
single byte. I capped the derived field numbers to 2047, which makes it
a little easier to think about encoded_len and avoids thinking about
reserved ranges or other limits.