Add deployment steps for derive crate (#425)
This builds on #324 to support a derive macro for `Encode`.

This PR sets up publishing for the `foxglove-derive` crate. 

The `bump-sdk-version` task will now keep the foxglove-derive dependency
pinned to the same version. This isn't strictly required, but seems
cleaner and avoids needing a manual bump (which is easy to forget since
we reference a path locally).

I moved the CI publish step to its own job. It now publishes the derive
crate, waits for the latest version to be available on crates.io, and
then publishes the foxglove crate as before.

As of right now, the dry-run CI check will still fail since we haven't
published the derive crate at all. I think once we publish a v0.7.1
version, that check will pass, and that's the easiest way forward. I
tested this with a local registry. We can publish that manually, or I
can publish derive 0.7.1 from a branch.

Lastly, I updated some of the derive crate metadata — notably, adding
the MIT license — and I moved some shared package config to the
workspace for clarity.

There are some third-party crates that can deal with this, but they
appear to address a lot of problems we have already solved, and at the
core still seem to do a sleep/poll for the dependent crate publishing.