C++: Support channel metadata (#479)
This adds support for channel metadata in C++. A user may supply
key/value pairs when constructing the channel.

The C library provides an iterator on top of the underlying BTreeMap,
which the C++ method uses to build up its own map. I expect metadata to
be typically small, and this getter to be called infrequently.

On the parameter order for `::create`: I added metadata last; I think it's possibly less common to use this than Context. If you want to supply metadata, but no Context, you have to provide the default `foxglove::Context()`. Longer-term, we could provide a builder interface around this if needed, but it was more than I wanted to tackle here.

I chose the generic struct name "FoxgloveKeyValue", thinking we might have other uses for that (e.g.,
supplying metadata for an MCAP file).