Containerize the C++ protobuf example (#494)
This is based on #483.

It adds a Dockerfile and updates the Readme for running the C++ protobuf
example. This example in particular takes some more effort to set up
(notably, the protobuf compiler), so this is the easiest way to see it
run. The example specifies running on linux/amd64 because it doesn't yet
support multi-arch builds; I think that's fine for seeing the example
work.

Tested on macOS (apple silicon).