Add C++ example showing logging custom generated protobuf (#483)
This adds a C++ example using a custom protobuf type generated with
google's C++ protobuf library.

It's painful, I think we could improve this story a lot by moving the
boilerplate for the schema and channel creation and serialization into
the Foxglove SDK.

The CMakeLists.txt needs some cleanup, it's probably not the minimal
version.

This also provides a Dockerfile and instructions to run the example in a
container.

---------

Co-authored-by: Bryan Fox <39674+bryfox@users.noreply.github.com>