Introduce strict build and configure CI to use it (#509)
### Changelog
None

### Docs
None

### Description
Enable a set of Clang and MSVC compiler warnings and set them to fail
builds if the `STRICT` compilation flag is set. This flag is set `true`
by default in containerized builds, including the ones that run as a
pre-merge check in GitHub Actions.

To disable these warnings, a user can run `make STRICT=false` from the
`cpp/` directory.

This PR also contains changes required to address the warnings in the
existing SDK codebase.

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

Builds fail if compiler warnings are enabled and escalated to errors

</td><td>

Builds pass with compiler warnings enabled

</td></tr></table>

Fixes: FG-12163

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->