Don't unadvertise a channel if an error prevented it from being advertised (#520)
### Changelog

Fixed a bug where a channel that fails to be advertised to a websocket
client would still be unadvertised. Now unadvertise is only sent if
advertise succeeded for a channel.

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

### Description


https://linear.app/foxglove/issue/FG-12137/sdk-websocket-server-should-not-unadvertise-channels-that-it-failed-to

I modified one of our tests to identify this issue and fail, then fixed
the bug and verified the test passes now.