Fix data loader seek from end (#529)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

None

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

None

### Description

The conversion to `SeekFrom::End` was around the wrong way. This came up
when trying to build an MCAP loader that seeks from the end of the file
to read summary info.