Thread through the client ID to the subscribe/unsubscribe callbacks for WebSocket server (#515)
### Changelog
- Create a new function, `Channel::log_to_sink`, allowing users to
optionally specify an individual Sink to target a logged message to. If
left `None`, the message is logged to all sinks associated with the
channel's context.
- Update C++ bindings for `Channel::log` to take a sink ID as an
optional argument.
- Update C++ `WebSocketServerCallbacks::onSubscribe` and
`WebSocketServerCallbacks::onUnsubscribe` function signatures to take
`foxglove::ClientMetadata` as a second parameter, which surfaces
information about which client and associated sinks are requesting to
subscribe or unsubscribe to a channel.

### Docs
Doxygen comments are updated.

### Description
These changes are motivated by work to port `ros-foxglove-bridge` to use
the SDK. The main problem being solved is ensuring that messages are
delivered to late-joining clients. The ability to target `log()` calls
to individual sinks gives WebSocketServer users more control over how
messages should be delivered in this situation (i.e. by implementing
logic that can republish a message to only a newly joining client
without sending a duplicate to others).

Surfacing information about which sink a given client is associated with
allows sinks to be used as the semantic for targeting message logging,
rather than coupling it to WebSocket client IDs.

Unit tests pass. Added additional tests to verify that messages are only
logged to the appropriate sinks if a target sink ID is passed.

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->
All calls to `Channel::log` deliver messages to every sink associated
with that channels's context.

</td><td>

<!--after content goes here-->
Calls to `Channel::log` can optionally be targeted to individual sinks.

</td></tr></table>

Fixes: FG-12285