Remove ws-protocol from public docs (#536)
This removes mention of "ws-protocol" from public documentation and
examples. ws-protocol is an implementation detail, and consumers of the
SDK shouldn't have to understand anything about it.

There are still internal Rust modules which link to the spec in many
places, which is still useful. (There was one part of the protocol
re-used in the public API, `Status`, which I updated.) I also added a CI
check to detect if generated docs accidentally re-introduce a link to
the protocol spec or source.