C++: use original `tl` namespace for expected (#543)
This updates `expected.hpp` to use the original `expected` header file
from https://github.com/TartanLlama/expected, to fix naming conflicts
when a project includes the same header. See #540.

Previously, when using our own namespace, we also conditionally (in
C++23) used `std::expected`. We can decide to re-introduce that later,
but for simplicity, let's use the header directly. I left the header in
the existing location, to avoid breaking any build setups (including
what we document in the quickstart).

To test this, I added the example here, which (at the first commit)
fails to build with the error described, but succeeds with the updated
header. I don't plan to merge this example; it's intended to help PR
review.

Fixes #540.