Introduce explicit integer Parameter type (#542)
### Changelog
- Introduce an explicit Integer (64-bit signed integer) type for
Parameters.
- Rename the Number parameter type to Float64 to make its semantic
clearer
- Update Rust/C/C++/Python bindings to allow constructing Integer and
IntegerArray parameters, and to rename Number to Float64 for
consistency.

### Docs
Documentation comments updated.

### Description
Previously, the Number parameter type was overloaded to represent both
integer and floating-point types (this is because when parameters are
sent over the wire as JSON, the distinction is lost). SDK users wanting
to construct integer parameters needed to cast them to doubles. When
these values were round-tripped through the WebSocketServer interface, a
fractional part of `.0` was added, which caused some deserialization
libraries to treat them as doubles.

While this can be worked around by SDK users by rolling their own
serialization and deserialization overloads to infer when a number
should be cast to a native integer type or kept as a double, it's more
ergonomic for the SDK to allow for constructing parameter values from
native integer types, and to make a best-effort attempt to serialize
integer parameters without adding a zero fractional part (ie.
`"my_param": 42` over `"my_param": 42.0`).

This PR also includes a custom JSON deserializer to more closely align
the behavior of parsing params from the wire into the behavior in the
protocol spec (namely, more closely respecting and enforcing checks
around the `type` tag).

<table><tr><th>Before</th><th>After</th></tr><tr><td>

Only a Number parameter type exists, which handles both floating point
and integer values.

</td><td>

A new Integer parameter type exists for handling explicit integers.
Number is deprecated and replaced with Float to be more explicit.

</td></tr></table>

Fixes: FG-12407