Fix typo in CameraCalibration docs (#553)
This fixes a typo in the description of the CameraCalibration message.

This also updates the `generate` task to build the C++ project as well,
since that in turn generates a C header from the rust source. The
contributing docs say to run this in a container, so it shouldn't break
the workflow if tooling is missing on the host.