Flesh out data loader docs, make backfill optional (#562)
### Changelog
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

None

### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

None

### Description

Fleshes out some of the comments on the structs in the WIT that make it
to the user but didn't have comments.

Also changes backfill to have an empty default implementation and more
comments.
Backfill isn't a strict requirement but does make playback better, so I
think having it as an advanced opt-in concept is better.

These docs will be published to docs.rs once the crate is released after
#559. So I'll try and get the updates to land before then.