Move test client headers to test directory, remove deps from bridge (#570)
### Changelog
None

### Docs
None
### Description

Move the test_client headers, used only by smoke_test to live in the
tests/ directory. This prevents them from being included by default in
the foxglove_bridge_component target, and allows that target to no
longer require linking with websocketpp and openssl.

Changed deps in the CMakeLists.txt and package.xml to reflect this.

Fixes: FG-12606

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->