Expose sink_id in channel logging functions (#567)
### Changelog
- [python] calls to `channel.log()` can now optionally specify a sink ID
to target message logging to a particular sink
- [c++] calls to `channel.log()` can now optionally specify a sink ID to
target message logging to a particular sink

### Docs
Docstrings for SDK documentation updated

### Description
Exposes an optional parameter for `log()` on channel bindings to target
logging messages to a specific sink. This is already being used
internally by `foxglove-bridge` to allow for selectively publishing
low-frequency/latched topics to late joining clients while avoiding
duplicate message delivery to already-connected clients.

Tests added to C++ server unit tests to test this behavior with an
actual WebSocket client connected in the test. I added a test to the
Python side to make sure that the parameter was wired through without
causing a runtime error. I held off on adding deeper behavior testing on
the Python side because it'd require a Python implementation of a
WebSocket client, and I wanted to keep the LoC count of this PR
tractable.

Fixes: FG-12408 FG-12405