Add `LocationFixes` schema; color `LocationFix` (#556)
Add a new `LocationFixes` schema for visualizing a series of fixes on
the Map panel, and a `Color` field to `LocationFix`

This adds a `Color` field to `LocationFix` and adds a new
`LocationFixes` schema. `LocationFixes` contains an array of
`LocationFix` messages.

There are a couple of use cases driving this change:

- A system tracking multiple entity locations publishes a
`LocationFixes` message containing a location fix for each entity at a
given timestamp. Each location's color can be used to differentiate when
visualized.
- Render a history of locations (track). Each fix has its own timestamp.

This array schema is reminiscent of the `FrameTransforms` schema.

Alternatives considered: Instead of adding Color, assign some sort of
group ID to location fix, which would allow the app to determine a color
at render time.