Package foxglove_bridge in docker image (#600)
### Changelog
- [Foxglove Bridge] The public beta of Foxglove Bridge is now packaged
as a Docker image in our public image registry

### Docs
README.md for bridge updated with image pull instructions

### Description
Create GHA workflow for pushing a Docker image containing the Foxglove
Bridge to our public image registry. This is in response to customer
feedback that providing these images helps them more easily deploy
Foxglove Bridge.

Images are built against all currently supported versions of ROS. The
GHA workflow is currently configured to build the bridge:
- On a push to `main`, where the version identifier will be the git SHA
of the commit on `main`
- On an SDK release tag (i.e. `sdk/vxxx`), where the version identifier
will be the SDK version number

The version number of Foxglove Bridge is incremented to equal that of
Foxglove SDK, as per our consensus that the versioning/release cadence
of these components should be coupled going forward.

Fixes: FLE-1