Support wss connections for bridge (#560)
This adds initial support for secure websocket connections (wss) in a
form that supports the ros-foxglove-bridge interface. Support is behind
an undocumented "tls" feature. We plan to expose this in the bridge
prior to SDK users in general. This does not contain any support for
Python.

The server accepts the contents of a PEM-formatted x509 certificate and
a PEM-formatted pkcs8 key. I tested this with a 4096-bit RSA key and
self-signed certificate, by trusting the cert in my keychain, using the
examples (which take file path arguments).

Under the hood, this uses rustls. Currently, it also uses
rustls-native-certs; TBD what we need to support here.