add encode method on all schemas (#614)
### Changelog
#### Added
- C: added `foxglove_..._encode` implementations for all foxglove
schemas (excluding Timestamp and Duration).
- C++: added `::encode()` methods to all foxglove schemas (excluding
Timestamp and Duration).
- Python: added `self.encode()` methods to all foxglove schemas
(excluding Timestamp and Duration).

These can be used to directly serialize the message as protobuf.
### Docs

None.

### Description

When attempting to use the foxglove SDK within a data loader, the user
needs to encode messages and return them to the caller, rather than pass
them to a Channel. The data loader framework code can't be wrapped in
such a way to make the Channel abstraction make sense, because messages
are "pulled" from data loaders rather than "pushed" into a channel.

Therefore, this PR adds encoding capability to C and C++ bindings (and
Python for completeness).

The C/C++ wrappers write into a preallocated buffer, and return an error
if the buffer is not large enough. This allows the caller to use
whatever allocation scheme suits them.

### Discussion

I'm not so sure about the language to use here in the documentation.
Should I explicitly say that these are encoded as protobuf? They use
whatever `Encode` implementation the underlying struct uses, which we
define for ourselves for our built-in schemas. Do we ever want to
implement other encodings for our own schemas and channels, or are we
happy to define that as Protobuf for now?

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->