C/C++ add schema() getter function for all builtin schemas (#616)
### Changelog
- C: add `foxglove_<schema name>_schema()` for Foxglove schemas
excluding Timestamp and Duration.
- C++: add `<schema name>::schema()` for Foxglove schemas excluding
Timestamp and Duration.

These allow the caller to retrieve runtime schema information that can
be used with MCAP or the Data Loader API.
<!-- Write a one-sentence summary of the user-impacting change (API,
UI/UX, performance, etc) that could appear in a changelog. Write "None"
if there is no user-facing change -->

### Docs

None

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

### Description

Following up from #614 , adds a getter for the schema information to the
C/C++ SDK. These mirror the similar `get_schema()` methods already
available in Python and Rust. This again allows data loaders to provide
schema information to the foxglove app in pull control flow rather than
push.
<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->