foxglove-c: enable building for WASM32 targets (#624)
### Changelog
- C: enable a minimal build for wasm32 targets, that includes Foxglove
schema struct definitions and serialization functionality only.
### Docs

<!-- Link to a Docs PR, tracking ticket in Linear, OR write "None" if no
documentation changes are needed. -->

### Description

We want to be able to use the foxglove SDK from within WASM data loaders
to serialize foxglove messages. To do this we need to build the core
foxglove serialization functionality as WASM.

This PR moves all of the code that does I/O in foxglove-c out of
`lib.rs` and into modules, which are conditionally included in the
library based on architecture. It uses cbindgen's `[defines]` feature to
put `#ifdef __wasm32__` around all of these features in the generated
header file.

NOTE: `__wasm32__` is not defined in any standard, but it is what the
WASI SDK Clang implementation uses, and that's the toolchain that we
specify for the data loader SDK.

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->