Support decoding of schemas & wkts (#627)
### Changelog
None

### Description

This adds a Decode trait with implementations for generated schemas, and
implements `merge_field` for decoding of Timestamp and Duration. This
will allow a client to decode existing messages without importing prost.
I think we could remove the explicit Decode trait (and only include the
merge_fields here) if we don't care about importing prost.

This is currently internal (docs hidden) for Agent work, and is not
exposed to Python/C++.