Fix UID conflict in Dockerfile.ros (#639)
### Changelog
None

### Docs
None

### Description
Fix issue in ROS Docker builds under Linux where the passed USER_UID
(typically 1000) conflicts with an existing user in the ROS base image.
Previously, this seems to have worked but after re-pulling the base
system layers, it seems a user now exists with UID 1000. This adds logic
to check if a user with that UID already exists and, if so, to use it.