Skip cloudflare deploy step if the PR is from an external fork. (#672)
### Changelog

None.

### Docs

None.

### Description

That's because the cloudflare deploy step requires secrets, which are
not passed to a job running on a PR from an external fork. By skipping
it, we lose a bit of convenience, but given that the number of PRs from
external forks is minimal, this should be OK.

You can see what happens with an external fork by looking at the CI in
https://github.com/foxglove/foxglove-sdk/pull/670 (ignore the CI/rust
failure, that's a different problem handled by #671).

Signed-off-by: Chris Lalancette <clalancette@gmail.com>