Update the workaround to find the service typesupport handle. (#673)
### Changelog

None.

### Docs

None.

### Description

The GenericClient implementation in foxglove_bridge needs to get a
service typesupport handle in order to initialize itself.

In late 2023, the rclcpp::get_service_typesupport_handle() API was added
to ROS as a one-call way to get the typesupport handle for services.
That means that all ROS 2 releases since Iron have that API.

Humble does not have that particular API. However, since late 2024, it
does include the underlying rosidl API, so we can call that without
resorting to hacks to find the symbol in the library. In Humble prior to
that 2024 release, we need that symbol lookup hack.

That leads to this patch. If we are on a new enough rclcpp, we use
rclcpp::get_service_typesupport_handle(). If we are on a newer version
of Humble, we use the underlying rosidl API.

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

All tests succeed (run with `colcon test`) on Kilted, Jazzy, Humble, and
older Humble (prior to the symbol being available).

Fixes FG-13130

---------

Signed-off-by: Chris Lalancette <chris.lalancette@foxglove.dev>