Add note to CameraCalibration that we do not support Tx and Ty values (#665)
Our `CameraCalibration` documentation makes it appear like we support
the `Tx` and `Ty` parameters of the projection matrix when in fact we
don't. This PR updates the documentation in the message to indicate we
do not currently support those parameters. We do leave the parameters
present so users can continue providing conventional projection matrix
values.