Ensure writer is closed in examples, add a note about that to doc string (#694)
### Changelog
None

### Docs

https://github.com/foxglove/docs/pull/1003/files

### Description

If you don't close the mcap writer, it won't finish the mcap file
correctly leading to an obscure error “length out of int32 range” when
you try to open it. So we add the with statement and a comment to all
examples, to make it clearer for the developer how to avoid that.