Rust: support filtering logging sinks by topic (#500)
This implements sink filtering, so that a user can (for example) log
"heavy" topics to a different MCAP than the rest of their data, but
still log all messages to a WS server in the same context. Python and
C++ will follow, but released at the same time.

The example and tests (which make up most of the changeset here)
illustrate the usage. I started with a version that also allowed you to
change filtering on the sink after the fact, but decided to keep it
simpler for now.