Python: support filtering logging sinks by topic (#504)
This builds on #500 to bring channel filters to Python.

This constructs a new channel descriptor when the filter is called, but
this isn't called frequently since we cache the results of this for the
MCAP sink.

The example is generally the same as the Rust one; I removed the point
cloud part of the live_visualization example in favor of this, since it
wasn't demonstrating much on its own there.

---------

Co-authored-by: Greg Smith <gasmith@foxglove.dev>