Update Docker image tagging scheme (#697)
### Changelog
- Foxglove Bridge Docker images now conform to the ROS package version
scheme

### Docs
Docs updated to reflect new Docker image tagging scheme

### Description
In #662 we updated the version of `foxglove_bridge` to match that of
`foxglove_msgs`, and we release it using the `ros-vx.x.x` git tag
(previously we had assumed the version of `foxglove_bridge` could be
tied to the SDK version + release instead). This updates our Docker
image publication and tagging logic to match.