JSONSchema: mark all properties required (#682)
### Changelog
- *Breaking*: Foxglove JSON schemas now require all object properties to
be defined. This updates the JSONSchema representation of Foxglove
schemas to match the requirements of the app, as well as the
requirements of ROS and Flatbuffer schemas.
- Note: The Foxglove app does not enforce this requirement at runtime,
so existing messages will not be validated differently within Foxglove.
### Docs

Self-documenting.
### Description

Today whether or not a field is required in a foxglove schema message
depends on the schema encoding you use. Today what the app can assume is
in a message depends on the encoding:
:white_check_mark:ROS(1|2)(msg|idl): all fields are required
❓ : Protobuf: no fields are required by the serializer. the deserializer
will fill in zero values for fields with primitive types. fields with
nested struct types are set to null by the deserializer.
:x:jsonschema: no fields are required.
:white_check_mark:Flatbuffer: all fields are required.

This PR updates our JSONSchemas to require fields, which brings us
closer to agreement between all encodings. It's impossible to require
fields in a protobuf schema (that's left to the application to enforce).

<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->