remove language referring to optional fields (#699)
### Changelog
None.

### Docs

Is docs

### Description

Foxglove schemas do not have a notion of optional fields, but some
docstrings refer to varying behavior based on "provided" values. This PR
updates the language to describe behavior of empty vs. non-empty values
in these cases. I've checked that the app behaves as described for
`LinePrimitive` and `TriangleListPrimtive` colors.
<!-- Describe the problem, what has changed, and motivation behind those
changes. Pretend you are advocating for this change and the reader is
skeptical. -->

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

<!--before content goes here-->

</td><td>

<!--after content goes here-->

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->