Update remote viz/teleop example (rust) (#700)
This replaces the rust "agent" example with one that produces a video
stream and accepts teleop messages. I'd like to include this. code (and
the python & C++ equivalents) in primary documentation, so I've tried to
keep LOC to a minimum while still illustrating both of those features.

The video feed is a two-color gradient which animates by sliding the x
offset. We have another example which produces image frames from a
camera, but takes more to get up and running (with a dependency on
OpenCV) and I'd like to avoid that here. Open to other ideas for the
visualization, though.